Internval argument can now be a float. Minor improvements and refactoring.

+++ cmd/gearman_top/main.go
@@ -18,12 +18,9 @@ func main() {
 	flag.BoolVar(&args.Q_quiet, "q", false, "Quiet mode")
 	flag.BoolVar(&args.B_batch, "b", false, "Batch mode")
 	flag.BoolVar(&args.V_verbose, "ver", false, "Verbose output")
-	flag.IntVar(&args.I_interval, "i", 1, "Set interval")
+	flag.Float64Var(&args.I_interval, "i", 1.0, "Set interval")
 	flag.Func("H", "Add host", modgearman.Add2HostList)
-
 	flag.Parse()
 
-	//modgearman.Send2gearmandadmin("status\nversion\n", "127.0.0.1", 4730)
-	//modgearman.PrintSatus()
 	modgearman.GearmanTop(&args)
 }
